refactor(ui): split handler.go into focused transport modules (#771)
Decomposes the 1216-line ui/handler.go into focused sibling files:
routes, auth, dashboard, objects, admin_ops, logs, responses. Adds
shared writeJSON/decodeJSON helpers and fixes writeJSONError to encode
the error message via json.NewEncoder so messages containing quotes or
backslashes produce valid JSON.

Closes #771
